refactor: move setup models to separate files for clarity
