+++ VirtualFinland.UsersAPI.Deployment/Common/Models/StackSetup.cs
@@ -1,0 +1,12 @@
+using Pulumi;
+
+namespace VirtualFinland.UsersAPI.Deployment.Common.Models;
+
+public record StackSetup
+{
+    public InputMap<string> Tags = default!;
+    public bool IsProductionEnvironment;
+    public string Environment = default!;
+    public string ProjectName = default!;
+    public VpcSetup VpcSetup = default!;
+}
