+++ VirtualFinland.UsersAPI.Deployment/Common/Models/VpcSetup.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using Pulumi;
+
+namespace VirtualFinland.UsersAPI.Deployment.Common.Models;
+
+public record VpcSetup
+{
+    public Input<string>? VpcId = default!;
+    public Output<IEnumerable<string>> PrivateSubnetIds = default!;
+}
