+++ VirtualFinland.UsersAPI.Deployment/UsersAPIStack.cs
@@ -3,6 +3,7 @@ using System.Collections.Immutable;
 using System.Linq;
 using Pulumi;
 using VirtualFinland.UsersAPI.Deployment.Common;
+using VirtualFinland.UsersAPI.Deployment.Common.Models;
 using VirtualFinland.UsersAPI.Deployment.Features;
 
 namespace VirtualFinland.UsersAPI.Deployment;
@@ -50,21 +51,6 @@ public class UsersApiStack : Stack
         var lambdaFunctionConfigs = new LambdaFunctionUrl(config, stackSetup, secretManagerSecret);
         ApplicationUrl = lambdaFunctionConfigs.ApplicationUrl;
         LambdaId = lambdaFunctionConfigs.LambdaFunctionArn;
-    }
-
-    public record StackSetup
-    {
-        public InputMap<string> Tags = default!;
-        public bool IsProductionEnvironment;
-        public string Environment = default!;
-        public string ProjectName = default!;
-        public VpcSetup VpcSetup = default!;
-    }
-
-    public record VpcSetup
-    {
-        public Input<string>? VpcId = default!;
-        public Output<IEnumerable<string>> PrivateSubnetIds = default!;
     }
 
     // Outputs for Pulumi service
